Terminado punto 4 del exament

+++ database/migrations/2019_06_17_091830_create_product_provider_table.php
@@ -6,11 +6,6 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreateProductProviderTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::create('product_provider', function (Blueprint $table) {
@@ -21,16 +16,11 @@ class CreateProductProviderTable extends Migration
            $table->unsignedBigInteger ('provider_id');
            $table->foreign ('provider_id')->references('id')->on('providers');
 
-           $table->integer ('price')->default(0);
+           $table->float ('price')->default(0);
            $table->integer ('discount')->default(0);
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::dropIfExists('product_provider');
